requirements.txt
pycryptodome   # is require if the program is run on Windows OS.